Extracting attachment management method to distinct files -> fixes #1940
