+++ browser/components/CodeEditor.js
@@ -3,7 +3,7 @@ import React from 'react'
 import _ from 'lodash'
 import CodeMirror from 'codemirror'
 import 'codemirror-mode-elixir'
-import attachmentManagement from 'browser/main/lib/dataApi/attachmentManagement'
+import attachmentManagement from 'browser/main/lib/attachmentManagement/index'
 import convertModeName from 'browser/lib/convertModeName'
 import eventEmitter from 'browser/main/lib/eventEmitter'
 import iconv from 'iconv-lite'
